+++ foo_ui_columns/replaygain_mode.cpp
@@ -52,7 +52,6 @@ struct ReplayGainModeToolbarArgs {
         auto playback_api = playback_control_v3::get();
         playback_api->restart();
     }
-    static const char* get_items_empty_text() { return nullptr; }
     static void get_menu_items(uie::menu_hook_t& p_hook)
     {
         p_hook.add_node(new cui::panel_helpers::CommandMenuNode{
@@ -75,6 +74,7 @@ struct ReplayGainModeToolbarArgs {
     static bool is_available() { return static_api_test_t<replaygain_manager_v2>(); }
     static std::unique_ptr<ReplayGainCoreSettingsNotifyLambda> callback;
     static constexpr bool refresh_on_click = false;
+    static constexpr auto no_items_text = ""sv;
     static constexpr const wchar_t* class_name{L"columns_ui_replaygain_mode_-bdvzDNKnwDniA"};
     static constexpr const char* name{"ReplayGain mode"};
     static constexpr GUID extension_guid{0xad9a81f7, 0x723a, 0x4cce, {0x87, 0xb6, 0x13, 0x39, 0xb, 0xda, 0xc2, 0x16}};