+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 * A new Output format toolbar was added, allowing the selection of the output bit depth for output devices that don’t use automatic output format selection. [[#389](https://github.com/reupen/columns_ui/pull/389), contributed by [@rplociennik](https://github.com/rplociennik)]
 
+* The DSP preset toolbar now displays the text '(no DSP presets exist)' if no DSP presets have been created. [[#395](https://github.com/reupen/columns_ui/pull/395)]
+
 ### Bug fixes
 
 * The minimum widths of the DSP preset and Output device toolbars now update if the list of DSP presets or output devices changes. [[#393](https://github.com/reupen/columns_ui/pull/393)]